refactor: club all the custom MDX Components inside mdxCompenents directory
