+++ src/content/concepts/output.mdx
@@ -9,7 +9,7 @@ contributors:
   - EugeneHlushko
 ---
 
-import CodeBlockWithCopy from "../../components/CodeBlockWithCopy/CodeBlockWithCopy";
+import CodeBlockWithCopy from "../../components/mdxComponents/CodeBlockWithCopy/CodeBlockWithCopy";
 
 Configuring the `output` configuration options tells webpack how to write the compiled files to disk. Note that, while there can be multiple `entry` points, only one `output` configuration is specified.
