+++ src/mdx-components.mjs
@@ -1,6 +1,6 @@
-import Badge from "./components/Badge/Badge.jsx";
-import CodeBlockWithCopy from "./components/CodeBlockWithCopy/CodeBlockWithCopy.jsx";
-import StackBlitzPreview from "./components/StackBlitzPreview/StackBlitzPreview.jsx";
+import Badge from "./components/mdxComponents/Badge/Badge.jsx";
+import CodeBlockWithCopy from "./components/mdxComponents/CodeBlockWithCopy/CodeBlockWithCopy.jsx";
+import StackBlitzPreview from "./components/mdxComponents/StackBlitzPreview/StackBlitzPreview.jsx";
 import LinkComponent from "./components/mdxComponents/Link.jsx";
 
 /** @returns {import('mdx/types.js').MDXComponents} */
